+++ app/controllers/comment_controller.rb
@@ -117,13 +117,13 @@ class CommentController < ApplicationController
 
   # Banned from adding comments?
   def reject_if_user_banned
-    return unless authenticated? && !authenticated_user.can_make_comments?
+    return if !authenticated? || authenticated_user.can_make_comments?
 
-    if authenticated_user.exceeded_limit?(:comments)
-      render template: 'comment/rate_limited'
-    else
+    if authenticated_user.banned?
       @details = authenticated_user.can_fail_html
       render template: 'user/banned'
+    else
+      render template: 'comment/rate_limited'
     end
   end